The issue of the side article not displaying has been resolved
